fix bug

+++ interactions/interactions.js
@@ -32,12 +32,15 @@ document.querySelector('#book-form').addEventListener('submit', (e) => {
 
 // Event: Remove a Book
 document.querySelector('#book-list').addEventListener('click', (e) => {
-    // Remove book from UI
-    UI.deleteBook(e.target);
+    // debugger
+    if (e.target.classList.contains('delete')) {
+        // Remove book from UI
+        UI.deleteBook(e.target);
 
-    // Remove book from store
-    Store.removeBook(e.target.parentElement.previousElementSibling.textContent);
+        // Remove book from store
+        Store.removeBook(e.target.parentElement.previousElementSibling.textContent);
 
-    // Show success message
-    UI.showAlert('Book Removed', 'success');
+        // Show success message
+        UI.showAlert('Book Removed', 'success');
+    }
 });
